Fixed an issue not checking if user is already part of a shortlist upon joining. -r1

+++ lib/sql/shortlist/is_member.sql
@@ -3,4 +3,4 @@ SELECT (
     FROM shortlists_users
     WHERE shortlist = $1 AND
     "user" = $2
-    ) AS is_member;
+    )::INT AS is_member;
